Chore(.module): move service package
