+++ src/test/java/com/newbarams/ajaja/module/auth/application/AuthMockBeans.java
@@ -1,4 +1,4 @@
-package com.newbarams.ajaja.module.auth.application.service;
+package com.newbarams.ajaja.module.auth.application;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.mock.mockito.MockBean;
