+++ src/test/java/com/newbarams/ajaja/module/user/application/ChangeReceiveTypeServiceTest.java
@@ -1,45 +1,44 @@
-package com.newbarams.ajaja.module.user.application.service;
+package com.newbarams.ajaja.module.user.application;
 
-import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 
 import com.newbarams.ajaja.common.support.MockTestSupport;
 import com.newbarams.ajaja.module.user.domain.Email;
-import com.newbarams.ajaja.module.user.domain.Nickname;
 import com.newbarams.ajaja.module.user.domain.User;
 import com.newbarams.ajaja.module.user.domain.UserRepository;
 
-class RenewNicknameServiceTest extends MockTestSupport {
+class ChangeReceiveTypeServiceTest extends MockTestSupport {
 	@InjectMocks
-	private RenewNicknameService renewNicknameService;
+	private ChangeReceiveTypeService changeReceiveTypeService;
 
 	@Mock
 	private RetrieveUserService retrieveUserService;
 	@Mock
 	private UserRepository userRepository;
 
-	@Test
-	@DisplayName("닉네임 갱신을 요청하면 새로운 이름으로 변경되어야 한다.")
-	void renew_Success_WithNewNickname() {
+	@ParameterizedTest
+	@EnumSource(User.ReceiveType.class)
+	@DisplayName("리마인드를 지원하는 종류로 변경할 수 있어야한다.")
+	void change_Success(User.ReceiveType type) {
 		// given
 		User user = sut.giveMeBuilder(User.class)
 			.set("email", new Email("Ajaja@me.com"))
 			.sample();
 
-		Nickname oldNickname = user.getNickname();
-		given(retrieveUserService.loadExistById(any())).willReturn(user);
+		given(retrieveUserService.loadExistById(anyLong())).willReturn(user);
 		given(userRepository.save(any())).willReturn(user);
 
 		// when
-		String newNickname = renewNicknameService.renew(user.getId());
+		changeReceiveTypeService.change(user.getId(), type);
 
 		// then
-		assertThat(oldNickname.getNickname()).isNotEqualTo(newNickname);
+		then(retrieveUserService).should(times(1)).loadExistById(anyLong());
 		then(userRepository).should(times(1)).save(any());
 	}
 }
