+++ src/test/java/com/newbarams/ajaja/module/user/application/RandomCertificationGeneratorTest.java
@@ -1,4 +1,4 @@
-package com.newbarams.ajaja.module.user.application.service;
+package com.newbarams.ajaja.module.user.application;
 
 import static org.assertj.core.api.Assertions.*;
